prepare for ios

+++ src/main/java/com/daxiang/core/ios/IosDevice.java
@@ -14,6 +14,6 @@ public class IosDevice extends MobileDevice {
 
     public int getMjpegServerPort() {
         Object mjpegServerPort = getAppiumDriver().getCapabilities().asMap().get("mjpegServerPort");
-        return (int) mjpegServerPort;
+        return (int)((long) mjpegServerPort);
     }
 }
